읽을 수 있게 — 글자 대비를 재서 올리고, SQL 을 절로 나눈다
글씨
  눈으로 «작다» 가 아니라 대비를 쟀다. --bg-surface 기준:
      text 15.4 OK · muted 6.0 OK · dim 3.1 큰글씨만 · faint 2.5 미달
  faint 는 본문(4.5)은 물론 큰 글씨(3.0)도 못 넘는데 시간·IP·축 눈금·힌트가
  전부 이 색이고 그것도 가장 작은 크기다. 작은 글씨 + 가장 낮은 대비가 겹친
  자리가 «안 보인다» 의 실체였다.

  호버 행은 배경이 밝아져 대비가 더 떨어진다 — 거기서도 4.5 를 넘게 잡았다:
      text 16.0/13.4 · muted 9.8/8.2 · dim 7.1/6.0 · faint 5.5/4.7
  그리고 micro 10px → 11px. 대비를 올려도 10px 은 10px 이다.

  Canvas 는 CSS 변수를 못 읽어 tokens.ts 에 값이 복제돼 있는데, 그 동기화를
  감시하는 테스트가 있어서 바로 잡혔다.

SQL
  프로파일의 SQL 은 한 줄로 온다. 150~400자짜리 join·서브쿼리가 벽처럼 붙어
  있어 어디가 where 인지 눈으로 못 찾는다. 여는 이유는 «어디가 느렸나» 인데
  문장을 못 읽으면 거기서 멈춘다.

  펼쳤을 때만 절 단위로 나눈다. 접힌 상태는 세 줄뿐이라 나누면 낱말 세 개만
  보인다 — 그때는 한 줄에 담는 편이 낫다.

  **글자를 하나도 바꾸지 않는다.** 공백을 접어 원문과 비교하는 테스트가
  그걸 지킨다. 실제로 두 번 잡혔다:
    · (select … 를 자르면서 없던 공백을 만들었다 → 공백이 있던 자리에서만 자른다
    · left outer join 이 «left / outer join» 으로 쪼개졌다 → 맨 앞도 알아보되
      줄은 안 나눈다
  따옴표·주석 안은 건드리지 않는다 (sqlBind 와 같은 규칙).

점검
  TS 568 · Rust 152 · 라이브 85 전부 통과. clippy 9건 중 7건 정리.
  남긴 2건은 고치면 뜻이 흐려지는 자리다.

+++ src-tauri/src/scouter/connection.rs
@@ -548,7 +548,7 @@ impl ScouterConnection {
     /// PerfCounterPack (Type 60) 파싱
     /// ASIS: PerfCounterPack.read(DataInputX)
     fn read_perf_counter_pack(&mut self) -> io::Result<PerfCounterPack> {
-        let time = self.read_decimal()? as i64;
+        let time = self.read_decimal()?;
         let obj_name = self.read_text()?;
         let timetype = self.read_byte()?;
